Fixes for cron taskmaster
